acl done

+++ backend/api/routes/acl.py
@@ -10,7 +10,7 @@ from api.models.db import Access, Secret, SecretOwners, Version
 from api.utils.cipher import AESCipher
 from api.utils.random_string import random_string
 
-from .models.access import ACLGet
+from .models.access import ACLGet, ACLPost
 
 
 acl = APIRouter(prefix="/secret/{name}/access")
@@ -36,7 +36,7 @@ async def get_acl(name: str, auth=Depends(UnionAuth(scopes=[]))):
 
 
 @acl.post(response_model=ACLGet)
-async def update_acl(name: str, new: ACLGet, auth=Depends(UnionAuth(scopes=[]))):
+async def update_acl(name: str, new: ACLPost, auth=Depends(UnionAuth(scopes=[]))):
     owner = (
         db.session.query(SecretOwners)
         .join(Secret)
@@ -45,16 +45,13 @@ async def update_acl(name: str, new: ACLGet, auth=Depends(UnionAuth(scopes=[])))
     )
     if not owner:
         raise HTTPException(status_code=404, detail="Secret not found")
-    if Access(owner.access) < Access("OWNER"):
+    if Access(owner.access) < Access("RW"):
         raise HTTPException(status_code=403, detail="ACL can be changed only by secret owner")
-    db.session.query(SecretOwners).join(Secret).filter(Secret.name == name).delete()
-    new_owner = 5
-
-    # result: dict[str, list[str] | str] = {}
-    # result["r"] = []
-    # result["rw"] = []
-    # for owner in owners:
-    #     if owner.owner_id == auth["id"]:
-    #         break
-    # else:
-    #     raise HTTPException(status_code=404, detail="Secret not found")
+    db.session.query(SecretOwners).join(Secret).filter(Secret.name == name, SecretOwners.access != "owner").delete()
+    for rw in new.rw:
+        db.session.add(SecretOwners(owner_id=rw, secret_id=owner.secret_id, access="rw"))
+    db.session.flush()
+    for r in new.r:
+        db.session.add(SecretOwners(owner_id=r, secret_id=owner.secret_id, access="rw"))
+    db.session.flush()
+    return {"owner": owner.owner_id, **new.model_dump()}
